requirements.txt
asgiref==3.8.1
attrs==24.2.0
CacheControl==0.14.1
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.0
Django==5.1.3
django-filter==24.3
django-multiselectfield==0.1.13
djangorestframework==3.15.2
drf-spectacular==0.27.2
firebase-admin==6.6.0
google-api-core==2.23.0
google-api-python-client==2.154.0
google-auth==2.36.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.19.0
google-cloud-storage==2.18.2
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.66.0
grpcio==1.68.0
grpcio-status==1.68.0
httplib2==0.22.0
idna==3.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
Markdown==3.7
MarkupSafe==3.0.2
msgpack==1.1.0
packaging==24.2
pillow==11.0.0
proto-plus==1.25.0
protobuf==5.29.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.10.1
pyparsing==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
requests==2.32.3
rpds-py==0.21.0
rsa==4.9
simplejson==3.19.3
sqlparse==0.5.2
uritemplate==4.1.1
urllib3==2.2.3
